chore: move BillingCounts to under the Billing context
