--- lib/logflare/queries/analytical_queries.ex
@@ -1,35 +1,0 @@
-defmodule Logflare.Queries.AnalyticalQueries do
-  import Ecto.Query
-
-  # e.g. https://elixirforum.com/t/help-understanding-difference-between-2-queries/41743/4
-
-  def wrap_in_generated_timeseries_subquery(
-        query,
-        group_by_column = :date,
-        aggregated_columns,
-        start_datetime,
-        end_datetime
-      )
-      when is_list(aggregated_columns) do
-    from(subquery(query), as: :t)
-    |> join(
-      :full,
-      [t: t],
-      t_nulls in fragment(
-        "SELECT generate_series(DATE(?), DATE(?), '1 day') as date",
-        ^start_datetime,
-        ^end_datetime
-      ),
-      on: field(t_nulls, ^group_by_column) == field(t, ^group_by_column),
-      as: :t_nulls
-    )
-    |> select([t: t, t_nulls: t_nulls], %{
-      date: coalesce(field(t, ^group_by_column), field(t_nulls, ^group_by_column))
-    })
-    |> select_merge([t: t], %{sum: coalesce(field(t, :sum), 0)})
-
-    # for c <- aggregated_columns, reduce: q do
-    #   q -> select_merge(q, [t: t], %{^c => coalesce(field(t, ^c), 0)})
-    # end
-  end
-end
