+++ lib/logflare/user.ex
@@ -78,7 +78,7 @@ defmodule Logflare.User do
     field :endpoints_beta, :boolean, default: false
     embeds_one :preferences, UserPreferences
 
-    has_many :billing_counts, Logflare.BillingCounts.BillingCount, on_delete: :delete_all
+    has_many :billing_counts, Logflare.Billing.BillingCount, on_delete: :delete_all
     has_many :sources, Source
     has_many :endpoint_queries, Logflare.Endpoint.Query
     has_many :vercel_auths, Vercel.Auth
